+++ public/userscripts/manga-tracker.user.js
@@ -41,8 +41,9 @@
 // @include      /^https?:\/\/mangazuki\.co\/read\/[a-zA-Z0-9_-]+\/[0-9\.]+$/
 // @include      /^https?:\/\/(reader\.)?ygscans\.com\/(reader\/)?read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^https?:\/\/reader\.championscans\.com\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
-// @updated      2017-08-30
-// @version      1.7.63
+// @include      /^http:\/\/puremashiro\.moe\/reader\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
+// @updated      2017-09-02
+// @version      1.7.64
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // @require      https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js
@@ -2335,6 +2336,17 @@ let sites = {
 	 * @type {SiteObject}
 	 */
 	'reader.championscans.com' : extendSite({
+		preInit : function(callback) {
+			this.setupFoolSlide();
+			callback();
+		}
+	}),
+
+	/**
+	 * PureMashiroScans (FoolSlide)
+	 * @type {SiteObject}
+	 */
+	'puremashiro.moe' : extendSite({
 		preInit : function(callback) {
 			this.setupFoolSlide();
 			callback();